requirements.txt
streamlit
yfinance
streamlit-autorefresh
pandas
gspread
google-auth